feat:change comment component injection to new design, inline component usage + creation ESPUB-16

+++ Frontend/src/app/modules/management-dialogs/node-comments/comments-list/comments-list.component.ts
@@ -1,36 +1,27 @@
-import {Component, Input, EventEmitter, Output, ViewChild, ElementRef, HostListener} from '@angular/core';
-import {RestConnectorService} from "../../../common/rest/services/rest-connector.service";
-import {Toast} from "../../../common/ui/toast";
-import {RestNodeService} from "../../../common/rest/services/rest-node.service";
-import {
-  NodeWrapper, Node, NodePermissions, LocalPermissionsResult, Permission,
-  LoginResult, UserProfile, Comments, Comment, User
-} from "../../../common/rest/data-object";
-import {ConfigurationService} from "../../../common/services/configuration.service";
-import {UIHelper} from "../../../common/ui/ui-helper";
-import {RestIamService} from "../../../common/rest/services/rest-iam.service";
-import {TranslateService} from "@ngx-translate/core";
-import {RestCommentsService} from "../../../common/rest/services/rest-comments.service";
-import {OptionItem} from "../../../common/ui/actionbar/option-item";
-import {RestConstants} from "../../../common/rest/rest-constants";
-import {DialogButton} from "../../../common/ui/modal-dialog/modal-dialog.component";
+import {DialogButton} from "../../../../common/ui/modal-dialog/modal-dialog.component";
+import {Comments, Comment, LoginResult, Node, User} from "../../../../common/rest/data-object";
+import {RestConstants} from "../../../../common/rest/rest-constants";
+import {OptionItem} from "../../../../common/ui/actionbar/option-item";
+import {RestNodeService} from "../../../../common/rest/services/rest-node.service";
 import {trigger} from "@angular/animations";
-import {UIAnimation} from "../../../common/ui/ui-animation";
+import {Component, EventEmitter, HostListener, Input, Output} from "@angular/core";
+import {RestConnectorService} from "../../../../common/rest/services/rest-connector.service";
+import {UIAnimation} from "../../../../common/ui/ui-animation";
+import {RestCommentsService} from "../../../../common/rest/services/rest-comments.service";
+import {RestIamService} from "../../../../common/rest/services/rest-iam.service";
+import {Toast} from "../../../../common/ui/toast";
 
 @Component({
-  selector: 'node-comments',
-  templateUrl: 'node-comments.component.html',
-  styleUrls: ['node-comments.component.scss'],
+  selector: 'comments-list',
+  templateUrl: 'comments-list.component.html',
+  styleUrls: ['comments-list.component.scss'],
   animations: [
     trigger('fade', UIAnimation.fade()),
     trigger('cardAnimation', UIAnimation.cardAnimation())
   ]
 })
-export class NodeCommentsComponent  {
+export class CommentsListComponent  {
   public _node: Node;
-  public dialogTitle:string;
-  public dialogMessage:string;
-  public dialogButtons:DialogButton[];
   private isGuest: boolean;
   private loading: boolean;
   private user: User;
@@ -54,15 +45,7 @@ export class NodeCommentsComponent  {
      */
   @Output() onChange=new EventEmitter();
 
-  @HostListener('document:keydown', ['$event'])
-  handleKeyboardEvent(event: KeyboardEvent) {
-      if(event.code=="Escape"){
-          event.preventDefault();
-          event.stopPropagation();
-          this.cancel();
-          return;
-      }
-  }
+
   constructor(
     private connector : RestConnectorService,
     private iam : RestIamService,
@@ -103,13 +86,11 @@ export class NodeCommentsComponent  {
     }
     if(isAuthor || this._node.access.indexOf(RestConstants.ACCESS_WRITE)!=-1){
       options.push(new OptionItem('NODE_COMMENTS.OPTION_DELETE','delete',()=>{
-        this.dialogTitle='NODE_COMMENTS.DELETE_COMMENT';
-        this.dialogMessage='NODE_COMMENTS.DELETE_COMMENT_MESSAGE';
-        this.dialogButtons=DialogButton.getYesNo(()=>{
-            this.dialogTitle=null;
+        this.toast.showModalDialog('NODE_COMMENTS.DELETE_COMMENT','NODE_COMMENTS.DELETE_COMMENT_MESSAGE',DialogButton.getYesNo(()=>{
+            this.toast.closeModalDialog()
           },()=>{
             this.onLoading.emit(true);
-            this.dialogTitle=null;
+            this.toast.closeModalDialog()
             this.commentsApi.deleteComment(comment.ref.id).subscribe(()=>{
               this.refresh();
               this.onChange.emit();
@@ -119,7 +100,7 @@ export class NodeCommentsComponent  {
               this.onLoading.emit(false);
             });
           }
-        );
+        ),true,false);
       }));
     }
     return options;
@@ -158,7 +139,7 @@ export class NodeCommentsComponent  {
       return;
     }
     this.commentsApi.getComments(this._node.ref.id).subscribe((data:Comments)=>{
-      this.comments=data.comments;
+      this.comments=data.comments.reverse();
       this.options=[];
       for(let comment of this.comments){
         this.options.push(this.getOptions(comment));
